Reorganize dag algorithms into new file
