+++ src/digraph.rs
@@ -50,9 +50,11 @@ use super::iterators::{
     EdgeIndexMap, EdgeIndices, EdgeList, NodeIndices, NodeMap, WeightedEdgeList,
 };
 use super::{
-    is_directed_acyclic_graph, DAGHasCycle, DAGWouldCycle, NoEdgeBetweenNodes,
-    NoSuitableNeighbors, NodesRemoved,
+    DAGHasCycle, DAGWouldCycle, NoEdgeBetweenNodes, NoSuitableNeighbors,
+    NodesRemoved,
 };
+
+use super::dag_algorithms::is_directed_acyclic_graph;
 
 /// A class for creating directed graphs
 ///
